rename database module
